🧹 Extract responsibilities from workhorse db __init__.py into separate files
