+++ packages/workhorse/src/settler_workhorse/db/exceptions.py
@@ -1,0 +1,12 @@
+"""Database layer exceptions."""
+
+class DatabaseError(Exception):
+    """Database operation error."""
+
+    pass
+
+
+class TenantContextError(Exception):
+    """Tenant context not set for RLS."""
+
+    pass
